services/vehicle-codifier/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vehicle-codifier"
version = "0.1.0"
description = "Unified vehicle codification service for AMIS CVEGS matching"
authors = ["Minca AI Team <team@mincaai.com>"]
packages = [{ include = "vehicle_codifier", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
# Web framework
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}

# Database
SQLAlchemy = "^2.0.30"
psycopg = {extras = ["binary"], version = "^3.1.18"}

# ML and similarity
rapidfuzz = "^3.9.0"
sentence-transformers = "^3.0.1"
openai = "^1.35.0"

# HTTP and async
httpx = "^0.27.0"
aioredis = "^2.0.1"

# Data processing
pandas = "^2.2.2"
numpy = "^1.26.4"
openpyxl = "^3.1.2"

# Utilities
pydantic = "^2.7.4"
pydantic-settings = "^2.3.0"
structlog = "^24.2.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"

# Shared packages
db = { path = "../../packages/db", develop = true }
ml = { path = "../../packages/ml", develop = true }
schemas = { path = "../../packages/schemas", develop = true }

[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
black = "^24.0"
ruff = "^0.5.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "TD", "FIX", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "PERF", "FURB", "LOG", "RUF"]
ignore = ["ANN101", "ANN102", "COM812", "ISC001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"